requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
numpy==2.4.2
opencv-python==4.13.0.92
requests==2.32.5
RPi.GPIO==0.7.1
screeninfo==0.8.1
urllib3==2.6.3
